added grand total

+++ POE_PART1/Purchase_calculations/Calculate_purchase.cs
@@ -11,13 +11,13 @@ namespace POE_PART1.Purchase_calculations
             _context = context;
         }
 
-        public async Task<float> GetTotalAmount()
+        public async Task<decimal> GetTotalAmount()
         {
             var Total_Money_Donations = await _context.Monetary_donations.SumAsync(m => m.Amount);
             var Total_Purchases = await _context.Purchase.SumAsync(g => g.Price);
 
             var TotalAmount = Total_Money_Donations - Total_Purchases;
-            return TotalAmount;
+            return (decimal)TotalAmount;
         }
 
     }
